+++ packages/shared/src/index.ts
@@ -80,3 +80,4 @@ export * from "./brokerIntegration.js";
 export * from "./marketing.js";
 export * from "./tenantIsolation.js";
 export * from "./crm.js";
+export * from "./emailIntel.js";